requirements.txt
torch
torchvision
timm
wandb
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
requests
streamlit
